backend/deepfake-backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
boto3==1.28.21
python-dotenv==1.0.0
tensorflow==2.16.1
transformers==4.30.2
opencv-python==4.8.0.76
pymongo==4.5.0
gunicorn==20.1.0